introduced new weight function by least loaded switch

+++ src/main/java/edu/asu/emit/algorithm/graph/BaseElementWithWeight.java
@@ -38,5 +38,5 @@ package edu.asu.emit.algorithm.graph;
  * @author snkas
  */
 public interface BaseElementWithWeight {
-	long getWeight();
+	double getWeight();
 }
